requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.3.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
pyjwt==2.8.0
cryptography==42.0.0
passlib[bcrypt]==1.7.4